# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple weather app API built with Go.",
        "title": "Weather App API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/weather/{city}": {
            "get": {
                "description": "Get weather data for a specified city",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Get weather data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City name",
                        "name": "city",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.WeatherResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.WeatherResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Weather description",
                    "type": "string"
                },
                "icon": {
                    "description": "Weather icon code",
                    "type": "string"
                },
                "temperature": {
                    "description": "Temperature in Fahrenheit",
                    "type": "number"
                }
            }
        }
    }
}